Docker: Allow overriding image names from .env
For a subset of the images required by a Compose or Swarm Setup,
non-default names can be configured in .env.  Originally, the names were
treated as the images' complete names, i.e., including an optional
registry domain.  Using this setup, it was possible to pull the
irregularly updated auxiliary images from a default registry while, at
the same time, obtaining certain images from a different registry.

Commit e225a57f97 changed this behavior.  Since then, the names in .env
can only be used to change part of the image name, excluding the
registry.  If a default registry is configured it is always prepended to
the given image name, breaking the above use case and compatibility with
older .env files.

This patch reintroduces (almost) full control over image names from
.env: if the given image name contains a slash, no additional registry
path is prepended.  At the same time it also continues to support the
new logic – debatable as it may be.