Docker: Simplify image name/registry handling
This patch removes the ability to override image names in .env.
Instead, only the registry of each image can be customized.

The reasoning here is that the only common reason to change an image
name is to change its Docker registry.  For example, while the default
registry may be set to default.example.com, it may be necessary to
obtain the backend image private.example.com/openslides-server.  With
this patch, this would be achieved by the following configuration in
.env:

    DOCKER_OPENSLIDES_BACKEND_REGISTRY="private.example.com"

For special cases, for which the rest of image name must indeed be
changed as well, the template would need to be customized.

The templates are not backwards-compatible.